aether mind alpha

+++ xRLAgents/agents/AgentAetherMindAlpha.py
@@ -220,8 +220,8 @@ class AgentAetherMindAlpha():
             self.optimizer.step() 
 
             # log results
-            self.log_loss_diffusion.add("mean", loss_diffusion.mean().detach().cpu().numpy())
-            self.log_loss_diffusion.add("std", loss_diffusion.std().detach().cpu().numpy())
+            self.log_loss_diffusion.add("mean", loss_diffusion.float().mean().detach().cpu().numpy())
+            self.log_loss_diffusion.add("std", loss_diffusion.float().std().detach().cpu().numpy())
                 
             for key in info_ssl:
                 self.log_loss_im_ssl.add(str(key), info_ssl[key])
